Openebs (#54)
fixes #52 by defaulting to flannel. There's a couple of other little enhancements (like the wait on ssh in the refresh), and lots of whitespace cleanup and a little bit of retabbing. Moving all the examples to their own folder as well. And abstracted out the net_urls necessary for kubectl apply -f for easier updating, they are now up top with the rest of the env vars.